# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory-API",
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://jeremybecker.ch/",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/inventoryitem": {
            "get": {
                "tags": [
                    "Inventory Item"
                ],
                "summary": "List inventory items",
                "description": "Get a List of all inventory items.",
                "operationId": "listInventoryItem",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Inventory Item"
                ],
                "summary": "Create inventory item",
                "description": "Create a new inventory item.",
                "operationId": "storeInventoryItem",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The name of the inventory item.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type_id",
                        "in": "path",
                        "description": "The id of the morph item.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type_type",
                        "in": "path",
                        "description": "The type of the morph item.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "The user id of the creator of the item.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "location_id",
                        "in": "path",
                        "description": "The id of the location the item is stored in.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/inventoryitem/{id}": {
            "get": {
                "tags": [
                    "Inventory Item"
                ],
                "summary": "Show inventory item",
                "description": "Display an inventory item",
                "operationId": "showInventoryItem",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Inventory Item"
                ],
                "summary": "Delete inventory item",
                "description": "Delete an inventory item.",
                "operationId": "deleteInventoryItem",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Inventory Item"
                ],
                "summary": "Update inventory item",
                "description": "Update variables in an inventory item.",
                "operationId": "updateInventoryItem",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The name of the inventory item.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type_id",
                        "in": "path",
                        "description": "The id of the morph item.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type_type",
                        "in": "path",
                        "description": "The type of the morph item.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "The user id of the creator of the item.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "location_id",
                        "in": "path",
                        "description": "The id of the location the item is stored in.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    }
}